EDP-195 refactor: seller product controller into `seller/` directory
